requirements.txt
python-dotenv>=1.0.0
openai>=1.30.0
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
sentence-transformers>=2.2.2
pypdf>=4.0.0
requests>=2.31.0
huggingface-hub>=0.22.0
torch>=2.0.0